// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  WORKER: 'WORKER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ContractStatus = {
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED'
} as const

export type ContractStatus = (typeof ContractStatus)[keyof typeof ContractStatus]


export const ActivityStatus = {
  UNASSIGNED: 'UNASSIGNED',
  SCHEDULED: 'SCHEDULED',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
  VERIFIED: 'VERIFIED',
  INVOICED: 'INVOICED'
} as const

export type ActivityStatus = (typeof ActivityStatus)[keyof typeof ActivityStatus]


export const InvoiceType = {
  CLIENT_BILL: 'CLIENT_BILL',
  WORKER_PAYOUT: 'WORKER_PAYOUT'
} as const

export type InvoiceType = (typeof InvoiceType)[keyof typeof InvoiceType]


export const InvoiceStatus = {
  DRAFT: 'DRAFT',
  SENT: 'SENT',
  PAID: 'PAID'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]